En déplaçant du code de ApplicationManager à MainActivity, le changement de langage se fait maintenant dans tout l'app, incluant le nav drawer! Par défaut (si l'utilisateur ne change pas de setting), l'app prend le langage par défaut du système
