+++ app/src/main/java/ca/etsmtl/applets/etsmobile/ApplicationManager.java
@@ -51,13 +51,9 @@ import io.supportkit.core.SupportKit;
 
 public class ApplicationManager extends Application {
 
-
-    public static LinkedHashMap<String, MyMenuItem> mMenu = new LinkedHashMap<String, MyMenuItem>(17);
     public static UserCredentials userCredentials;
     public static String domaine;
     public static int typeUsagerId;
-    public SharedPreferences prefs;
-    public Locale locale;
 
     @Override
     protected void attachBaseContext(Context base) {
@@ -70,14 +66,9 @@ public class ApplicationManager extends Application {
         super.onCreate();
         AnalyticsHelper.getInstance(this);
         createDatabaseTables();
-        /*prefs = this.getSharedPreferences("Language", 0);
-        locale = new Locale(prefs.getString("language","fr"));
-        Locale.setDefault(locale);
-        Log.d("AppManager", "language: " + locale.toString());*/
 
         SupportKit.init(this, getString(R.string.credentials_supportkit));
         Fabric.with(this, new Crashlytics());
-        setTitles();
 
         AccountManager accountManager = AccountManager.get(this);
         Account[] accounts = accountManager.getAccountsByType(Constants.ACCOUNT_TYPE);
@@ -103,156 +94,9 @@ public class ApplicationManager extends Application {
 
     }
 
-    public void setTitles(){
-        // Section 1 - Moi
-        mMenu.put(getString(R.string.menu_section_1_moi),
-                new MyMenuItem(
-                        getString(R.string.menu_section_1_moi),
-                        null
-                ));
-
-        mMenu.put(TodayFragment.class.getName(),
-                new MyMenuItem(
-                        getString(R.string.menu_section_1_ajd),
-                        TodayFragment.class,
-                        R.drawable.ic_ico_aujourdhui,
-                        true
-                ));
-
-        mMenu.put(HoraireFragment.class.getName(),
-                new MyMenuItem(
-                        getString(R.string.menu_section_1_horaire),
-                        HoraireFragment.class,
-                        R.drawable.ic_ico_schedule,
-                        true
-                ));
-
-        mMenu.put(NotesFragment.class.getName(),
-                new MyMenuItem(
-                        getString(R.string.menu_section_1_notes),
-                        NotesFragment.class,
-                        R.drawable.ic_ico_notes,
-                        true
-                ));
-
-        mMenu.put(MoodleFragment.class.getName(),
-                new MyMenuItem(
-                        getString(R.string.menu_section_2_moodle),
-                        MoodleFragment.class,
-                        R.drawable.ic_moodle_icon_small,
-                        true
-                ));
-
-        mMenu.put(ProfilFragment.class.getName(),
-                new MyMenuItem(
-                        getString(R.string.menu_section_1_profil),
-                        ProfilFragment.class,
-                        R.drawable.ic_ico_profil,
-                        true
-                ));
-
-        mMenu.put(BandwithFragment.class.getName(),
-                new MyMenuItem(
-                        getString(R.string.menu_section_1_bandwith),
-                        BandwithFragment.class,
-                        R.drawable.ic_ico_internet,
-                        false
-                ));
-
-
-        // Section 2 - ÉTS
-        mMenu.put(getString(R.string.menu_section_2_ets),
-                new MyMenuItem(
-                        getString(R.string.menu_section_2_ets),
-                        null
-                ));
-
-        mMenu.put(NewsFragment.class.getName(),
-                new MyMenuItem(
-                        getString(R.string.menu_section_2_news),
-                        NewsFragment.class,
-                        R.drawable.ic_ico_news,
-                        false
-                ));
-
-
-        mMenu.put(BottinFragment.class.getName(),
-                new MyMenuItem(
-                        getString(R.string.menu_section_2_bottin),
-                        BottinFragment.class,
-                        R.drawable.ic_ico_bottin,
-                        false
-                ));
-
-        mMenu.put(BiblioFragment.class.getName(),
-                new MyMenuItem(
-                        getString(R.string.menu_section_2_biblio),
-                        BiblioFragment.class,
-                        R.drawable.ic_ico_library,
-                        false
-                ));
-
-        mMenu.put(SecuriteFragment.class.getName(),
-                new MyMenuItem(
-                        getString(R.string.menu_section_2_securite),
-                        SecuriteFragment.class,
-                        R.drawable.ic_ico_security,
-                        false
-                ));
-
-        // Section 3 - ApplETS
-        mMenu.put(getString(R.string.menu_section_3_applets),
-                new MyMenuItem(
-                        getString(R.string.menu_section_3_applets),
-                        null
-                ));
-
-        mMenu.put(OtherAppsFragment.class.getName(),
-                new MyMenuItem(
-                        getString(R.string.menu_section_3_apps),
-                        OtherAppsFragment.class,
-                        R.drawable.ic_star_60x60,
-                        false
-                ));
-
-        mMenu.put(AboutFragment.class.getName(),
-                new MyMenuItem(
-                        getString(R.string.menu_section_3_about),
-                        AboutFragment.class,
-                        R.drawable.ic_logo_icon_final,
-                        false
-                ));
-
-        mMenu.put(CommentairesFragment.class.getName(),
-                new MyMenuItem(
-                        getString(R.string.menu_section_3_comms),
-                        CommentairesFragment.class,
-                        R.drawable.ic_ico_comment,
-                        false
-                ));
-
-        mMenu.put(SponsorsFragment.class.getName(),
-                new MyMenuItem(
-                        getString(R.string.menu_section_3_sponsors),
-                        SponsorsFragment.class,
-                        R.drawable.ic_ico_partners,
-                        false
-                ));
-
-        mMenu.put(FAQFragment.class.getName(),
-                new MyMenuItem(
-                        getString(R.string.menu_section_3_faq),
-                        FAQFragment.class,
-                        R.drawable.ic_ico_faq,
-                        false
-                ));
-    }
-
     @Override
     public void onConfigurationChanged(Configuration newConfig) {
         super.onConfigurationChanged(newConfig);
-        setTitles();
-        Log.d("AppManager", "language: " + locale.toString());
     }
 
     public static void deconnexion(final Activity activity) {
